Added GM calculation
The submerged volume from the function combineVolumes(array) in volumeCalculations.js, line 35 returns a volume under z (waterline) = 0.
I do not know if this is intentional, but it seems that this is caused because some volumes (in "calculations") are <0 and they counter each other.

I've  added a cheat line to avoid V being 0 that calculates V as disp = Lpp * B * T *Cb.