Multibody examples, introduced shipState.objectOverrides.baseByGroup, sketched new ShipState
The Jensen multibody example does not work very well yet.

ShipState_new.js is not tested or in use. It just points out a possible direction for future development. In fact, I think a new ShipState must be very different, if the goal is to have efficient, seamless caching (when necessary) and good integration with Ship3D and other potential users of the class. We will probably have to use Proxy objects and Object.defineProperty with custom getters/setters. Maybe we will even want to have a ShipSimulation class that connects a Ship with a ShipState.